Added all types of unit tests

+++ StringManager/StringManager.Core/Models/InstalledString.cs
@@ -22,8 +22,8 @@ namespace StringManager.Core.Models
 
         public string ToneName { get; set; }
 
-        public int Frequency { get; set; }
+        public double Frequency { get; set; }
 
-        public int WaveLenght { get; set; }
+        public double WaveLenght { get; set; }
     }
 }
